requirements.txt
plotly==4.14.3
streamlit==0.76.0
tensorflow-cpu==2.3.0
numpy>=1.18.0
requests==2.24.0
Pillow==8.1.0
pandas==1.1.1